refactor(agentic): improve code health and complexity across tools and tests
